refactor: organize components structure
