+++ app/layout.tsx
@@ -3,7 +3,7 @@ import localFont from 'next/font/local'
 import './globals.css'
 
 import { Toaster } from '@/components/ui/toaster'
-import Footer from '@/components/ui/Footer'
+import Footer from '@/components/Footer'
 
 const geistSans = localFont({
   src: './fonts/GeistVF.woff',
